refactor(#2971): remove duplicated variables
