+++ tests/integration/test_fim/test_files/test_report_changes/test_file_size_default.py
@@ -64,12 +64,12 @@ import os
 
 import pytest
 from test_fim.common import generate_string, translate_size, make_diff_file_path
-from wazuh_testing import global_parameters, LOG_FILE_PATH
+from wazuh_testing import global_parameters, LOG_FILE_PATH, REGULAR
 from wazuh_testing.tools import PREFIX
 from wazuh_testing.tools.file import create_file, modify_file_content
 from wazuh_testing.tools.configuration import load_wazuh_configurations
 from wazuh_testing.tools.monitoring import FileMonitor, generate_monitoring_callback
-from wazuh_testing.modules.fim import (REGULAR, CB_FILE_SIZE_LIMIT_REACHED, FIM_DEFAULT_LOCAL_INTERNAL_OPTIONS,
+from wazuh_testing.modules.fim import (CB_FILE_SIZE_LIMIT_REACHED, FIM_DEFAULT_LOCAL_INTERNAL_OPTIONS,
                                        ERR_MSG_FIM_EVENT_NOT_DETECTED, ERR_MSG_FILE_LIMIT_REACHED)
 from wazuh_testing.modules.fim.event_monitor import callback_detect_event
 from wazuh_testing.modules.fim.utils import generate_params
