+++ tests/integration/test_fim/test_files/test_report_changes/test_file_size_disabled.py
@@ -66,8 +66,8 @@ import pytest
 from wazuh_testing.tools import PREFIX
 from wazuh_testing.tools.configuration import load_wazuh_configurations
 from wazuh_testing.tools.monitoring import FileMonitor, generate_monitoring_callback
-from wazuh_testing import global_parameters, LOG_FILE_PATH
-from wazuh_testing.modules.fim import REGULAR, CB_FILE_SIZE_LIMIT_REACHED, FIM_DEFAULT_LOCAL_INTERNAL_OPTIONS
+from wazuh_testing import global_parameters, LOG_FILE_PATH, REGULAR
+from wazuh_testing.modules.fim import CB_FILE_SIZE_LIMIT_REACHED, FIM_DEFAULT_LOCAL_INTERNAL_OPTIONS
 from wazuh_testing.modules.fim.utils import generate_params, create_file
 from test_fim.common import generate_string
